Extract AbstractMvcEndpoint class
- for implementing MvcEndpoints that don't have a backing Endpoint
