+++ spring-boot-actuator/src/main/java/org/springframework/boot/actuate/endpoint/mvc/DocsMvcEndpoint.java
@@ -16,15 +16,10 @@
 
 package org.springframework.boot.actuate.endpoint.mvc;
 
-import org.springframework.boot.actuate.endpoint.Endpoint;
-import org.springframework.boot.actuate.endpoint.EndpointProperties;
 import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.context.EnvironmentAware;
-import org.springframework.core.env.Environment;
 import org.springframework.http.MediaType;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.servlet.config.annotation.ResourceHandlerRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurerAdapter;
 
 /**
  * {@link MvcEndpoint} to expose actuator documentation.
@@ -33,93 +28,41 @@ import org.springframework.web.servlet.config.annotation.WebMvcConfigurerAdapter
  * @since 1.3.0
  */
 @ConfigurationProperties("endpoints.docs")
-public class DocsMvcEndpoint extends WebMvcConfigurerAdapter
-		implements MvcEndpoint, EnvironmentAware {
+public class DocsMvcEndpoint extends AbstractMvcEndpoint {
 
 	private static final String DOCS_LOCATION = "classpath:/META-INF/resources/spring-boot-actuator/docs/";
-
-	private Environment environment;
-
-	/**
-	 * Endpoint URL path.
-	 */
-	private String path = "/docs";
-
-	/**
-	 * Enable the endpoint.
-	 */
-	private boolean enabled = true;
-
-	/**
-	 * Mark if the endpoint exposes sensitive information.
-	 */
-	private Boolean sensitive;
 
 	private final ManagementServletContext managementServletContext;
 
 	private Curies curies = new Curies();
-
-	@Override
-	public void setEnvironment(Environment environment) {
-		this.environment = environment;
-	}
 
 	public Curies getCuries() {
 		return this.curies;
 	}
 
 	public DocsMvcEndpoint(ManagementServletContext managementServletContext) {
+		setPath("/docs");
+		this.sensitiveDefault = false;
 		this.managementServletContext = managementServletContext;
 	}
 
 	@RequestMapping(value = "/", produces = MediaType.TEXT_HTML_VALUE)
 	public String browse() {
-		return "forward:" + this.managementServletContext.getContextPath() + this.path
+		return "forward:" + this.managementServletContext.getContextPath() + getPath()
 				+ "/index.html";
 	}
 
 	@RequestMapping(value = "", produces = MediaType.TEXT_HTML_VALUE)
 	public String redirect() {
-		return "redirect:" + this.managementServletContext.getContextPath() + this.path
+		return "redirect:" + this.managementServletContext.getContextPath() + getPath()
 				+ "/";
 	}
 
 	@Override
 	public void addResourceHandlers(ResourceHandlerRegistry registry) {
 		registry.addResourceHandler(
-				this.managementServletContext.getContextPath() + this.path + "/**")
+				this.managementServletContext.getContextPath() + getPath() + "/**")
 				.addResourceLocations(DOCS_LOCATION);
-	}
-
-	public void setPath(String path) {
-		this.path = path;
-	}
-
-	@Override
-	public String getPath() {
-		return this.path;
-	}
-
-	public boolean isEnabled() {
-		return this.enabled;
-	}
-
-	public void setEnabled(boolean enabled) {
-		this.enabled = enabled;
-	}
-
-	@Override
-	public boolean isSensitive() {
-		return EndpointProperties.isSensitive(this.environment, this.sensitive, false);
-	}
-
-	public void setSensitive(Boolean sensitive) {
-		this.sensitive = sensitive;
-	}
-
-	@Override
-	public Class<? extends Endpoint<?>> getEndpointType() {
-		return null;
 	}
 
 	/**
