+++ spring-boot-actuator/src/main/java/org/springframework/boot/actuate/endpoint/mvc/HalJsonMvcEndpoint.java
@@ -16,20 +16,12 @@
 
 package org.springframework.boot.actuate.endpoint.mvc;
 
-import javax.validation.constraints.NotNull;
-import javax.validation.constraints.Pattern;
-
-import org.springframework.boot.actuate.endpoint.Endpoint;
-import org.springframework.boot.actuate.endpoint.EndpointProperties;
 import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.context.EnvironmentAware;
-import org.springframework.core.env.Environment;
 import org.springframework.hateoas.ResourceSupport;
 import org.springframework.http.MediaType;
 import org.springframework.util.StringUtils;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseBody;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurerAdapter;
 
 /**
  * {@link MvcEndpoint} to expose HAL-formatted JSON.
@@ -40,38 +32,13 @@ import org.springframework.web.servlet.config.annotation.WebMvcConfigurerAdapter
  * @since 1.3.0
  */
 @ConfigurationProperties("endpoints.actuator")
-public class HalJsonMvcEndpoint extends WebMvcConfigurerAdapter
-		implements MvcEndpoint, EnvironmentAware {
-
-	private Environment environment;
-
-	/**
-	 * Endpoint URL path.
-	 */
-	@NotNull
-	@Pattern(regexp = "^$|/.*", message = "Path must be empty or start with /")
-	private String path;
-
-	/**
-	 * Enable the endpoint.
-	 */
-	private boolean enabled = true;
-
-	/**
-	 * Mark if the endpoint exposes sensitive information.
-	 */
-	private Boolean sensitive;
-
+public class HalJsonMvcEndpoint extends AbstractMvcEndpoint {
 	private final ManagementServletContext managementServletContext;
 
 	public HalJsonMvcEndpoint(ManagementServletContext managementServletContext) {
 		this.managementServletContext = managementServletContext;
-		this.path = getDefaultPath(managementServletContext);
-	}
-
-	@Override
-	public void setEnvironment(Environment environment) {
-		this.environment = environment;
+		this.sensitiveDefault = false;
+		setPath(getDefaultPath(managementServletContext));
 	}
 
 	private String getDefaultPath(ManagementServletContext managementServletContext) {
@@ -87,39 +54,7 @@ public class HalJsonMvcEndpoint extends WebMvcConfigurerAdapter
 		return new ResourceSupport();
 	}
 
-	public void setPath(String path) {
-		this.path = path;
-	}
-
-	@Override
-	public String getPath() {
-		return this.path;
-	}
-
-	public boolean isEnabled() {
-		return this.enabled;
-	}
-
-	public void setEnabled(boolean enabled) {
-		this.enabled = enabled;
-	}
-
-	@Override
-	public boolean isSensitive() {
-		return EndpointProperties.isSensitive(this.environment, this.sensitive, false);
-	}
-
-	public void setSensitive(Boolean sensitive) {
-		this.sensitive = sensitive;
-	}
-
-	@Override
-	public Class<? extends Endpoint<?>> getEndpointType() {
-		return null;
-	}
-
 	protected final ManagementServletContext getManagementServletContext() {
 		return this.managementServletContext;
 	}
-
 }
